Added template-based description.
I realized I didn't put my email and phone number. I then added them in
the same template as the rest of the CV.
I had to modify a bit the url writing as well as the other tags in the
template to make it look better and more modular.

I also have added meta information.