Corrected clicking issue.
The container with the `minimized` toggle embraced everything so we
couldn't click on anything inside. This bug was introduced when modifying
the template to add the label just before the URL.

Corrected by changing the behaviour of `url` by adding `before` to put
text before the link.

Also added forgotten file cv-description.js.